Commented out CSV related functions
It's been a while since I worked on this code but my belief is that these functions and tests are no longer in use in the app due to the removal of all CSV use.
My assumption was correct. Will remove these functions during a wider refactor in progress currently.